requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.1
djangorestframework==3.14.0
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pytest==8.0.0
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
